header as seperate component use of ion-router-outlet
