+++ ui/src/app/about/about.component.html
@@ -1,31 +1,34 @@
-<ion-grid>
-  <ion-row class="ion-justify-content-center">
-    <ion-col size="12" size-md="8">
-      <ion-card>
-        <ion-item lines="full" color="light">
-          <ion-icon slot="start" size="large" color="primary" name="information-circle-outline"></ion-icon>
-          <ion-label class="ion-text-wrap">
-            OpenEMS UI
-            <ion-card-subtitle translate>About.ui</ion-card-subtitle>
-          </ion-label>
-        </ion-item>
+<header></header>
+<ion-content>
+  <ion-grid>
+    <ion-row class="ion-justify-content-center">
+      <ion-col size="12" size-md="8">
+        <ion-card>
+          <ion-item lines="full" color="light">
+            <ion-icon slot="start" size="large" color="primary" name="information-circle-outline"></ion-icon>
+            <ion-label class="ion-text-wrap">
+              OpenEMS UI
+              <ion-card-subtitle translate>About.ui</ion-card-subtitle>
+            </ion-label>
+          </ion-item>
 
-        <ion-card-content class="ion-text-wrap">
-          <p translate>About.developed</p>
-          <ul>
-            <li>
-              <a target="_blank" href="https://openems.io" translate>About.openEMS</a>
-            </li>
-            <li>
-              <a target="_blank" href="https://github.com/OpenEMS/openems/tree/develop">
-                <span translate>About.build</span>: 2020.18.0-SNAPSHOT</a>
-            </li>
-          </ul>
-          <p>
-            <span [innerHTML]="'About.contact' | translate:{value: 'info@openems.io'}"></span>
-          </p>
-        </ion-card-content>
-      </ion-card>
-    </ion-col>
-  </ion-row>
-</ion-grid>
+          <ion-card-content class="ion-text-wrap">
+            <p translate>About.developed</p>
+            <ul>
+              <li>
+                <a target="_blank" href="https://openems.io" translate>About.openEMS</a>
+              </li>
+              <li>
+                <a target="_blank" href="https://github.com/OpenEMS/openems/tree/develop">
+                  <span translate>About.build</span>: 2020.18.0-SNAPSHOT</a>
+              </li>
+            </ul>
+            <p>
+              <span [innerHTML]="'About.contact' | translate:{value: 'info@openems.io'}"></span>
+            </p>
+          </ion-card-content>
+        </ion-card>
+      </ion-col>
+    </ion-row>
+  </ion-grid>
+</ion-content>
