+++ ui/src/app/edge/history/autarchy/autarchychartoverview/autarchychartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge">
+<ng-container *ngIf="edge">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title translate>General.autarchy</ion-title>
@@ -14,13 +14,15 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ion-card-content>
-        <table class="full_width">
-            <tr>
-                <td style="width: 100%" class="ion-padding"></td>
-            </tr>
-        </table>
-        <autarchychart [period]="service.historyPeriod">
-        </autarchychart>
-    </ion-card-content>
-</ion-content>
+    <ion-content>
+        <ion-card-content>
+            <table class="full_width">
+                <tr>
+                    <td style="width: 100%" class="ion-padding"></td>
+                </tr>
+            </table>
+            <autarchychart [period]="service.historyPeriod">
+            </autarchychart>
+        </ion-card-content>
+    </ion-content>
+</ng-container>
