+++ ui/src/app/edge/history/channelthreshold/channelthresholdchartoverview/channelthresholdchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge && config">
+<ng-container *ngIf="edge && config">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title>{{ component.alias }}</ion-title>
@@ -17,26 +17,28 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ng-container *ngIf="showTotal">
-        <ion-card-content>
-            <table class="full_width">
-                <tr>
-                    <td style="width: 100%" class="ion-padding" translate>General.total</td>
-                </tr>
-            </table>
-            <channelthresholdTotalChart [period]="service.historyPeriod">
-            </channelthresholdTotalChart>
-        </ion-card-content>
-    </ng-container>
-    <ng-container *ngIf="!showTotal || showTotal == null">
-        <ion-card-content>
-            <table class="full_width">
-                <tr>
-                    <td style="width: 100%" class="ion-padding"></td>
-                </tr>
-            </table>
-            <channelthresholdSingleChart [componentId]="component.id" [period]="service.historyPeriod">
-            </channelthresholdSingleChart>
-        </ion-card-content>
-    </ng-container>
-</ion-content>
+    <ion-content>
+        <ng-container *ngIf="showTotal">
+            <ion-card-content>
+                <table class="full_width">
+                    <tr>
+                        <td style="width: 100%" class="ion-padding" translate>General.total</td>
+                    </tr>
+                </table>
+                <channelthresholdTotalChart [period]="service.historyPeriod">
+                </channelthresholdTotalChart>
+            </ion-card-content>
+        </ng-container>
+        <ng-container *ngIf="!showTotal || showTotal == null">
+            <ion-card-content>
+                <table class="full_width">
+                    <tr>
+                        <td style="width: 100%" class="ion-padding"></td>
+                    </tr>
+                </table>
+                <channelthresholdSingleChart [componentId]="component.id" [period]="service.historyPeriod">
+                </channelthresholdSingleChart>
+            </ion-card-content>
+        </ng-container>
+    </ion-content>
+</ng-container>
