+++ ui/src/app/edge/history/chpsoc/chpsocchartoverview/chpsocchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge && config">
+<ng-container *ngIf="edge && config">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title translate>{{ component.alias }}</ion-title>
@@ -14,8 +14,10 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ion-card-content>
-        <chpsocchart [componentId]="component.id" [period]="service.historyPeriod">
-        </chpsocchart>
-    </ion-card-content>
-</ion-content>
+    <ion-content>
+        <ion-card-content>
+            <chpsocchart [componentId]="component.id" [period]="service.historyPeriod">
+            </chpsocchart>
+        </ion-card-content>
+    </ion-content>
+</ng-container>
