+++ ui/src/app/edge/history/fixdigitaloutput/fixdigitaloutputchartoverview/fixdigitaloutputchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge">
+<ng-container *ngIf="edge">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title>{{ component.alias }}</ion-title>
@@ -17,26 +17,28 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ng-container *ngIf="showTotal">
-        <ion-card-content>
-            <table class="full_width">
-                <tr>
-                    <td style="width: 100%" class="ion-padding" translate>General.total</td>
-                </tr>
-            </table>
-            <fixDigitalOutputTotalChart [period]="service.historyPeriod">
-            </fixDigitalOutputTotalChart>
-        </ion-card-content>
-    </ng-container>
-    <ng-container *ngIf="!showTotal || showTotal == null">
-        <ion-card-content>
-            <table class="full_width">
-                <tr>
-                    <td style="width: 100%" class="ion-padding"></td>
-                </tr>
-            </table>
-            <fixDigitalOutputSingleChart [componentId]="component.id" [period]="service.historyPeriod">
-            </fixDigitalOutputSingleChart>
-        </ion-card-content>
-    </ng-container>
-</ion-content>
+    <ion-content>
+        <ng-container *ngIf="showTotal">
+            <ion-card-content>
+                <table class="full_width">
+                    <tr>
+                        <td style="width: 100%" class="ion-padding" translate>General.total</td>
+                    </tr>
+                </table>
+                <fixDigitalOutputTotalChart [period]="service.historyPeriod">
+                </fixDigitalOutputTotalChart>
+            </ion-card-content>
+        </ng-container>
+        <ng-container *ngIf="!showTotal || showTotal == null">
+            <ion-card-content>
+                <table class="full_width">
+                    <tr>
+                        <td style="width: 100%" class="ion-padding"></td>
+                    </tr>
+                </table>
+                <fixDigitalOutputSingleChart [componentId]="component.id" [period]="service.historyPeriod">
+                </fixDigitalOutputSingleChart>
+            </ion-card-content>
+        </ng-container>
+    </ion-content>
+</ng-container>
