+++ ui/src/app/edge/history/grid/gridchartoverview/gridchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge">
+<ng-container *ngIf="edge">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title translate>General.grid</ion-title>
@@ -17,25 +17,27 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ion-card-content class="underline">
-        <table class="full_width">
-            <tr>
-                <td style="width: 100%" class="ion-padding"></td>
-            </tr>
-        </table>
-        <gridChart [showPhases]="showPhases" [period]="service.historyPeriod"></gridChart>
-    </ion-card-content>
-    <!-- Info Text -->
-    <ion-card-content class="underline">
-        <ion-row class="ion-justify-content-center">
-            <ion-item class="ion-text-center" lines="none">
-                <ion-icon color="primary" size="large" slot="start" name="alert-outline"></ion-icon>
-                <ion-label class="ion-text-wrap">
-                    <small translate>
-                        Edge.Index.Widgets.twoWayInfoGrid
-                    </small>
-                </ion-label>
-            </ion-item>
-        </ion-row>
-    </ion-card-content>
-</ion-content>
+    <ion-content>
+        <ion-card-content class="underline">
+            <table class="full_width">
+                <tr>
+                    <td style="width: 100%" class="ion-padding"></td>
+                </tr>
+            </table>
+            <gridChart [showPhases]="showPhases" [period]="service.historyPeriod"></gridChart>
+        </ion-card-content>
+        <!-- Info Text -->
+        <ion-card-content class="underline">
+            <ion-row class="ion-justify-content-center">
+                <ion-item class="ion-text-center" lines="none">
+                    <ion-icon color="primary" size="large" slot="start" name="alert-outline"></ion-icon>
+                    <ion-label class="ion-text-wrap">
+                        <small translate>
+                            Edge.Index.Widgets.twoWayInfoGrid
+                        </small>
+                    </ion-label>
+                </ion-item>
+            </ion-row>
+        </ion-card-content>
+    </ion-content>
+</ng-container>
