+++ ui/src/app/edge/history/peakshaving/symmetric/symmetricpeakshavingchartoverview/symmetricpeakshavingchart.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge && component">
+<ng-container *ngIf="edge && component">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title>{{ component.alias }}</ion-title>
@@ -14,8 +14,10 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ion-card-content>
-        <symmetricpeakshavingchart [componentId]="component.id" [period]="service.historyPeriod">
-        </symmetricpeakshavingchart>
-    </ion-card-content>
-</ion-content>
+    <ion-content>
+        <ion-card-content>
+            <symmetricpeakshavingchart [componentId]="component.id" [period]="service.historyPeriod">
+            </symmetricpeakshavingchart>
+        </ion-card-content>
+    </ion-content>
+</ng-container>
