+++ ui/src/app/edge/history/selfconsumption/selfconsumptionchartoverview/selfconsumptionchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge">
+<ng-container *ngIf="edge">
     <ion-header>
         <ion-toolbar class="ion-justify-content-center" color="primary">
             <ion-title class="ion-text-start ion-padding-left" translate>General.selfConsumption</ion-title>
@@ -14,12 +14,14 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ion-card-content>
-        <table class="full_width">
-            <tr>
-                <td style="width: 100%" class="ion-padding"></td>
-            </tr>
-        </table>
-        <selfconsumptionChart [period]="service.historyPeriod"></selfconsumptionChart>
-    </ion-card-content>
-</ion-content>
+    <ion-content>
+        <ion-card-content>
+            <table class="full_width">
+                <tr>
+                    <td style="width: 100%" class="ion-padding"></td>
+                </tr>
+            </table>
+            <selfconsumptionChart [period]="service.historyPeriod"></selfconsumptionChart>
+        </ion-card-content>
+    </ion-content>
+</ng-container>
