+++ ui/src/app/edge/history/singlethreshold/singlethresholdchartoverview/singlethresholdchartoverview.component.html
@@ -1,4 +1,4 @@
-<ion-content *ngIf="edge && component">
+<ng-container *ngIf="edge && component">
     <ion-header>
         <ion-toolbar color="primary">
             <ion-title>{{ component.alias }}</ion-title>
@@ -14,28 +14,30 @@
             </ion-buttons>
         </ion-toolbar>
     </ion-header>
-    <ng-container>
-        <ion-card-content>
-            <table class="full_width">
-                <tr>
-                    <td style="width: 100%" class="ion-padding"></td>
-                </tr>
-            </table>
-            <singlethresholdChart [componentId]="component.id" [period]="service.historyPeriod">
-            </singlethresholdChart>
+    <ion-content>
+        <ng-container>
+            <ion-card-content>
+                <table class="full_width">
+                    <tr>
+                        <td style="width: 100%" class="ion-padding"></td>
+                    </tr>
+                </table>
+                <singlethresholdChart [componentId]="component.id" [period]="service.historyPeriod">
+                </singlethresholdChart>
+            </ion-card-content>
+        </ng-container>
+        <!-- Info Text -->
+        <ion-card-content class="underline" *ngIf="inputChannel == '_sum/GridActivePower'">
+            <ion-row class="ion-justify-content-center">
+                <ion-item class="ion-text-center" lines="none">
+                    <ion-icon color="primary" size="large" slot="start" name="alert-outline"></ion-icon>
+                    <ion-label class="ion-text-wrap">
+                        <small translate>
+                            Edge.Index.Widgets.twoWayInfoGrid
+                        </small>
+                    </ion-label>
+                </ion-item>
+            </ion-row>
         </ion-card-content>
-    </ng-container>
-    <!-- Info Text -->
-    <ion-card-content class="underline" *ngIf="inputChannel == '_sum/GridActivePower'">
-        <ion-row class="ion-justify-content-center">
-            <ion-item class="ion-text-center" lines="none">
-                <ion-icon color="primary" size="large" slot="start" name="alert-outline"></ion-icon>
-                <ion-label class="ion-text-wrap">
-                    <small translate>
-                        Edge.Index.Widgets.twoWayInfoGrid
-                    </small>
-                </ion-label>
-            </ion-item>
-        </ion-row>
-    </ion-card-content>
-</ion-content>
+    </ion-content>
+</ng-container>
