+++ ui/src/app/edge/settings/component/install/install.component.html
@@ -1,25 +1,27 @@
-<ion-grid *ngIf="factory && form">
-  <ion-row class="ion-justify-content-center">
+<header></header>
+<ion-content>
+  <ion-grid *ngIf="factory && form">
+    <ion-row class="ion-justify-content-center">
+      <ion-col size="12" size-xl="8">
+        <ion-card>
+          <ion-item>
+            <ion-icon slot="start" name="stats-outline" color="primary"></ion-icon>
+            <ion-label>
+              <ion-card-title class="ion-text-wrap">{{ factory.name }}</ion-card-title>
+              <ion-card-subtitle class="ion-text-wrap">{{ factory.description }}</ion-card-subtitle>
+            </ion-label>
+          </ion-item>
+          <ion-card-content>
+            <form [formGroup]="form" (ngSubmit)="submit()">
+              <formly-form [form]="form" [fields]="fields" [model]="model"></formly-form>
+              <ion-item>
+                <ion-button type="submit" color="primary">Komponente anlegen</ion-button>
+              </ion-item>
+            </form>
+          </ion-card-content>
+        </ion-card>
+      </ion-col>
 
-    <ion-col size="12" size-xl="8">
-      <ion-card>
-        <ion-item>
-          <ion-icon slot="start" name="stats-outline" color="primary"></ion-icon>
-          <ion-label>
-            <ion-card-title class="ion-text-wrap">{{ factory.name }}</ion-card-title>
-            <ion-card-subtitle class="ion-text-wrap">{{ factory.description }}</ion-card-subtitle>
-          </ion-label>
-        </ion-item>
-        <ion-card-content>
-          <form [formGroup]="form" (ngSubmit)="submit()">
-            <formly-form [form]="form" [fields]="fields" [model]="model"></formly-form>
-            <ion-item>
-              <ion-button type="submit" color="primary">Komponente anlegen</ion-button>
-            </ion-item>
-          </form>
-        </ion-card-content>
-      </ion-card>
-    </ion-col>
-
-  </ion-row>
-</ion-grid>
+    </ion-row>
+  </ion-grid>
+</ion-content>
