+++ ui/src/app/edge/settings/component/update/index.component.html
@@ -1,27 +1,30 @@
-<ion-searchbar (ionChange)="updateFilter($event.detail.value)"></ion-searchbar>
+<header></header>
+<ion-content>
+  <ion-searchbar (ionChange)="updateFilter($event.detail.value)"></ion-searchbar>
 
-<ion-list>
-  <ng-container *ngFor="let entry of list">
-    <ion-item-group>
-      <ion-item lines="inset" *ngIf="entry.filteredComponents.length>= 1">
-        <ion-icon name="{{entry.category.icon}}"></ion-icon>
-        <ion-label class="ion-padding-start">
-          {{ entry.category.title }}
+  <ion-list>
+    <ng-container *ngFor="let entry of list">
+      <ion-item-group>
+        <ion-item lines="inset" *ngIf="entry.filteredComponents.length>= 1">
+          <ion-icon name="{{entry.category.icon}}"></ion-icon>
+          <ion-label class="ion-padding-start">
+            {{ entry.category.title }}
+          </ion-label>
+        </ion-item>
+      </ion-item-group>
+      <ion-item lines="none" *ngFor="let item of entry.filteredComponents" [routerLink]=item.id>
+        <ion-label style="cursor: pointer" class="ion-text-wrap">
+          <ion-text color="primary">
+            <h3>{{ item.alias }}
+              <span *ngIf="item.id != item.alias"> ({{ item.id }})</span>
+              <small *ngIf="!item.isEnabled"> (<span translate>General.componentInactive</span>)</small>
+            </h3>
+          </ion-text>
+          <ng-container *ngIf="config.factories[item.factoryId] as factory">
+            <p>{{ factory.name }} <small>({{ factory.description }})</small></p>
+          </ng-container>
         </ion-label>
       </ion-item>
-    </ion-item-group>
-    <ion-item lines="none" *ngFor="let item of entry.filteredComponents" [routerLink]=item.id>
-      <ion-label style="cursor: pointer" class="ion-text-wrap">
-        <ion-text color="primary">
-          <h3>{{ item.alias }}
-            <span *ngIf="item.id != item.alias"> ({{ item.id }})</span>
-            <small *ngIf="!item.isEnabled"> (<span translate>General.componentInactive</span>)</small>
-          </h3>
-        </ion-text>
-        <ng-container *ngIf="config.factories[item.factoryId] as factory">
-          <p>{{ factory.name }} <small>({{ factory.description }})</small></p>
-        </ng-container>
-      </ion-label>
-    </ion-item>
-  </ng-container>
-</ion-list>
+    </ng-container>
+  </ion-list>
+</ion-content>
