+++ ui/src/app/edge/settings/systemlog/systemlog.component.html
@@ -1,41 +1,45 @@
-<ion-grid>
-  <ion-row>
-    <ion-col size="12">
-      <ion-card>
-        <ion-item lines="full" color="light">
-          <ion-icon slot="start" name="code-working-outline" color="primary"></ion-icon>
-          <ion-label>System-Log</ion-label>
-        </ion-item>
-
-        <ion-card-content>
-          <ion-item>
-            <ion-checkbox (ionChange)="toggleSubscribe($event)" slot="start" [(ngModel)]="isSubscribed"></ion-checkbox>
-            <ion-label translate class="ion-text-wrap">Edge.Config.Log.automaticUpdating</ion-label>
+<header></header>
+<ion-content>
+  <ion-grid>
+    <ion-row>
+      <ion-col size="12">
+        <ion-card>
+          <ion-item lines="full" color="light">
+            <ion-icon slot="start" name="code-working-outline" color="primary"></ion-icon>
+            <ion-label>System-Log</ion-label>
           </ion-item>
-        </ion-card-content>
 
-        <ion-card-content>
-          <ion-grid *ngFor="let line of lines" class="ion-no-padding">
-            <ion-row style="font-family: monospace; font-size: small" style.color="black">
-              <ion-col size-sm="2" size="6">
-                {{ line.time }}
-              </ion-col>
-              <ion-col size-sm="1" size="6">
-                {{ line.level }}
-              </ion-col>
-              <ion-col size-sm="auto" size="12">
-                {{ line.source }}
-              </ion-col>
-            </ion-row>
-            <ion-row style="font-family: monospace" [style.color]="line.color">
-              <ion-col size="12">
-                {{ line.message }}
-              </ion-col>
-            </ion-row>
-          </ion-grid>
-        </ion-card-content>
+          <ion-card-content>
+            <ion-item>
+              <ion-checkbox (ionChange)="toggleSubscribe($event)" slot="start" [(ngModel)]="isSubscribed">
+              </ion-checkbox>
+              <ion-label translate class="ion-text-wrap">Edge.Config.Log.automaticUpdating</ion-label>
+            </ion-item>
+          </ion-card-content>
 
-      </ion-card>
-    </ion-col>
-  </ion-row>
-</ion-grid>
+          <ion-card-content>
+            <ion-grid *ngFor="let line of lines" class="ion-no-padding">
+              <ion-row style="font-family: monospace; font-size: small" style.color="black">
+                <ion-col size-sm="2" size="6">
+                  {{ line.time }}
+                </ion-col>
+                <ion-col size-sm="1" size="6">
+                  {{ line.level }}
+                </ion-col>
+                <ion-col size-sm="auto" size="12">
+                  {{ line.source }}
+                </ion-col>
+              </ion-row>
+              <ion-row style="font-family: monospace" [style.color]="line.color">
+                <ion-col size="12">
+                  {{ line.message }}
+                </ion-col>
+              </ion-row>
+            </ion-grid>
+          </ion-card-content>
+
+        </ion-card>
+      </ion-col>
+    </ion-row>
+  </ion-grid>
+</ion-content>
