+++ ui/src/app/index/index.component.html
@@ -1,95 +1,98 @@
-<ion-grid>
-  <ion-row class="ion-justify-content-center">
-    <ion-col size="12" size-md="6" size-lg="4">
+<header></header>
+<ion-content>
+  <ion-grid>
+    <ion-row class="ion-justify-content-center">
+      <ion-col size="12" size-md="6" size-lg="4">
 
-      <!-- Waiting for authentication -->
-      <ion-card *ngIf="env.backend === 'OpenEMS Edge' && websocket.status == 'waiting for authentication'">
-        <ion-item color="primary" lines="full">
-          <ion-icon slot="start" name="contact-outline"></ion-icon>
-          <ion-label>Login</ion-label>
-        </ion-item>
-
-        <ion-card-content>
-          <form (ngSubmit)="doLogin(password.value)">
-            <div class="ion-padding">
-              <ion-item>
-                <ion-text class="ion-text-wrap" color="medium">
-                  Bitte geben Sie Ihr Passwort ein oder bestätigen Sie die Voreingabe um sich als Gast anzumelden.
-                </ion-text>
-              </ion-item>
-              <ion-item>
-                <ion-label position="floating">Passwort</ion-label>
-                <ion-input #password required type="password" placeholder="Passwort" value="user"></ion-input>
-              </ion-item>
-              <ion-button class="ion-float-right" type="submit">
-                <ion-icon slot="icon-only" name="log-in-outline"></ion-icon>
-              </ion-button>
-            </div>
-          </form>
-        </ion-card-content>
-
-      </ion-card>
-    </ion-col>
-  </ion-row>
-
-  <ng-container *ngIf="env.backend === 'OpenEMS Backend' && websocket.status == 'online'">
-
-    <!--
-      No access to Edges
-    -->
-    <ion-row *ngIf="noEdges; else showEdges" class="ion-justify-content-center">
-      <ion-col size-md="8" size-xl="4">
-        <ion-card>
-          <ion-item color="warning" lines="full">
-            <ion-icon slot="start" name="alert-circle-outline"></ion-icon>
-            <ion-label>No access</ion-label>
+        <!-- Waiting for authentication -->
+        <ion-card *ngIf="env.backend === 'OpenEMS Edge' && websocket.status == 'waiting for authentication'">
+          <ion-item color="primary" lines="full">
+            <ion-icon slot="start" name="contact-outline"></ion-icon>
+            <ion-label>Login</ion-label>
           </ion-item>
 
           <ion-card-content>
-            <p>Unforatunately you seem to not have permission to access any OpenEMS Edge device.</p>
+            <form (ngSubmit)="doLogin(password.value)">
+              <div class="ion-padding">
+                <ion-item>
+                  <ion-text class="ion-text-wrap" color="medium">
+                    Bitte geben Sie Ihr Passwort ein oder bestätigen Sie die Voreingabe um sich als Gast anzumelden.
+                  </ion-text>
+                </ion-item>
+                <ion-item>
+                  <ion-label position="floating">Passwort</ion-label>
+                  <ion-input #password required type="password" placeholder="Passwort" value="user"></ion-input>
+                </ion-item>
+                <ion-button class="ion-float-right" type="submit">
+                  <ion-icon slot="icon-only" name="log-in-outline"></ion-icon>
+                </ion-button>
+              </div>
+            </form>
           </ion-card-content>
+
         </ion-card>
       </ion-col>
     </ion-row>
 
-    <!--
-      Show searchbar and list of Edges
+    <ng-container *ngIf="env.backend === 'OpenEMS Backend' && websocket.status == 'online'">
+
+      <!--
+      No access to Edges
     -->
-    <ng-template #showEdges>
-      <ion-row class="ion-justify-content-center">
-        <ion-col size-md="12" size-xl="4">
-          <ion-searchbar [(ngModel)]="filter" (ionChange)="updateFilteredEdges()"></ion-searchbar>
-        </ion-col>
-      </ion-row>
-
-      <ion-row class="ion-justify-content-center">
-        <ion-col size-md="12" size-xl="4">
-          <ion-list lines="full">
-
-            <ion-item *ngFor="let edge of filteredEdges | slice:0:slice; let i=index" lines="inset"
-              [routerLink]="['/device', edge.id]" [color]="edge.isOnline ? null : 'light'">
-              <ion-label>
-                <h2 class="ion-text-wrap">{{ edge.comment }}</h2>
-                <p *ngIf="!edge.isOnline" class="ion-text-wrap">Das Gerät ist nicht verbunden!</p>
-              </ion-label>
-              <ion-label>
-                <ng-container *ngIf="env.backend === 'OpenEMS Backend'">
-                  <p>ID: {{ edge.id }}</p>
-                  <p class="ion-text-wrap">Typ: {{ edge.producttype }}</p>
-                </ng-container>
-                <p class="ion-text-wrap">Angemeldet als: {{ edge.getRoleString() }}.</p>
-              </ion-label>
-              <ion-icon slot="end" name="play-outline" color="primary"></ion-icon>
+      <ion-row *ngIf="noEdges; else showEdges" class="ion-justify-content-center">
+        <ion-col size-md="8" size-xl="4">
+          <ion-card>
+            <ion-item color="warning" lines="full">
+              <ion-icon slot="start" name="alert-circle-outline"></ion-icon>
+              <ion-label>No access</ion-label>
             </ion-item>
 
-            <ion-infinite-scroll (ionInfinite)="doInfinite($event)">
-              <ion-infinite-scroll-content loadingSpinner="bubbles" distance="1%" loadingText="Loading more data...">
-              </ion-infinite-scroll-content>
-            </ion-infinite-scroll>
-
-          </ion-list>
+            <ion-card-content>
+              <p>Unforatunately you seem to not have permission to access any OpenEMS Edge device.</p>
+            </ion-card-content>
+          </ion-card>
         </ion-col>
       </ion-row>
-    </ng-template>
-  </ng-container>
-</ion-grid>
+
+      <!--
+      Show searchbar and list of Edges
+    -->
+      <ng-template #showEdges>
+        <ion-row class="ion-justify-content-center">
+          <ion-col size-md="12" size-xl="4">
+            <ion-searchbar [(ngModel)]="filter" (ionChange)="updateFilteredEdges()"></ion-searchbar>
+          </ion-col>
+        </ion-row>
+
+        <ion-row class="ion-justify-content-center">
+          <ion-col size-md="12" size-xl="4">
+            <ion-list lines="full">
+
+              <ion-item *ngFor="let edge of filteredEdges | slice:0:slice; let i=index" lines="inset"
+                [routerLink]="['/device', edge.id]" [color]="edge.isOnline ? null : 'light'">
+                <ion-label>
+                  <h2 class="ion-text-wrap">{{ edge.comment }}</h2>
+                  <p *ngIf="!edge.isOnline" class="ion-text-wrap">Das Gerät ist nicht verbunden!</p>
+                </ion-label>
+                <ion-label>
+                  <ng-container *ngIf="env.backend === 'OpenEMS Backend'">
+                    <p>ID: {{ edge.id }}</p>
+                    <p class="ion-text-wrap">Typ: {{ edge.producttype }}</p>
+                  </ng-container>
+                  <p class="ion-text-wrap">Angemeldet als: {{ edge.getRoleString() }}.</p>
+                </ion-label>
+                <ion-icon slot="end" name="play-outline" color="primary"></ion-icon>
+              </ion-item>
+
+              <ion-infinite-scroll (ionInfinite)="doInfinite($event)">
+                <ion-infinite-scroll-content loadingSpinner="bubbles" distance="1%" loadingText="Loading more data...">
+                </ion-infinite-scroll-content>
+              </ion-infinite-scroll>
+
+            </ion-list>
+          </ion-col>
+        </ion-row>
+      </ng-template>
+    </ng-container>
+  </ion-grid>
+</ion-content>
