+++ ui/src/app/settings/settings.component.html
@@ -1,37 +1,42 @@
-<ion-grid>
-    <ion-row class="ion-justify-content-center">
-        <ion-col size-xl="6" size-xs="12">
-            <ion-card>
-                <ion-item lines="full" color="light">
-                    <ion-icon slot="start" size="large" color="primary" name="options-outline"></ion-icon>
-                    <ion-label translate>Menu.generalSettings</ion-label>
-                </ion-item>
-                <ion-card-content>
-                    <ion-item>
-                        <ion-label translate>About.language</ion-label>
-                        <ion-select #langSelect (ionChange)="setLanguage(langSelect.value)" [value]="currentLanguage">
-                            <ion-select-option *ngFor="let language of languages" [value]="language">
-                                {{ language }}
-                            </ion-select-option>
-                        </ion-select>
+<header></header>
+<ion-content>
+    <ion-grid>
+        <ion-row class="ion-justify-content-center">
+            <ion-col size-xl="6" size-xs="12">
+                <ion-card>
+                    <ion-item lines="full" color="light">
+                        <ion-icon slot="start" size="large" color="primary" name="options-outline"></ion-icon>
+                        <ion-label translate>Menu.generalSettings</ion-label>
                     </ion-item>
-                </ion-card-content>
-            </ion-card>
-        </ion-col>
-        <ion-col size-xl="6" size-xs="12">
-            <ion-card>
-                <ion-item lines="full" color="light">
-                    <ion-icon slot="start" size="large" color="primary" name="build-outline"></ion-icon>
-                    <ion-label>Debug Mode</ion-label>
-                </ion-item>
-                <ion-card-content>
-                    <ion-item>
-                        <ion-checkbox (ionChange)="toggleDebugMode($event)" slot="start" [(ngModel)]="env.debugMode">
-                        </ion-checkbox>
-                        <ion-label translate>Debug-Mode</ion-label>
+                    <ion-card-content>
+                        <ion-item>
+                            <ion-label translate>About.language</ion-label>
+                            <ion-select #langSelect (ionChange)="setLanguage(langSelect.value)"
+                                [value]="currentLanguage">
+                                <ion-select-option *ngFor="let language of languages" [value]="language">
+                                    {{ language }}
+                                </ion-select-option>
+                            </ion-select>
+                        </ion-item>
+                    </ion-card-content>
+                </ion-card>
+            </ion-col>
+            <ion-col size-xl="6" size-xs="12">
+                <ion-card>
+                    <ion-item lines="full" color="light">
+                        <ion-icon slot="start" size="large" color="primary" name="build-outline"></ion-icon>
+                        <ion-label>Debug Mode</ion-label>
                     </ion-item>
-                </ion-card-content>
-            </ion-card>
-        </ion-col>
-    </ion-row>
-</ion-grid>
+                    <ion-card-content>
+                        <ion-item>
+                            <ion-checkbox (ionChange)="toggleDebugMode($event)" slot="start"
+                                [(ngModel)]="env.debugMode">
+                            </ion-checkbox>
+                            <ion-label translate>Debug-Mode</ion-label>
+                        </ion-item>
+                    </ion-card-content>
+                </ion-card>
+            </ion-col>
+        </ion-row>
+    </ion-grid>
+</ion-content>
