+++ ui/src/app/shared/header/header.component.html
@@ -1,0 +1,70 @@
+<ion-header>
+  <ion-toolbar color="primary" mode="md">
+    <ion-buttons *ngIf="backUrl" slot="start">
+      <ion-button [routerLink]="backUrl">
+        <ion-icon name="arrow-back-outline"></ion-icon>
+      </ion-button>
+    </ion-buttons>
+    <ion-buttons slot="start" *ngIf="enableSideMenu">
+      <ion-menu-button autoHide="false"></ion-menu-button>
+    </ion-buttons>
+    <!-- page title is shown if the current page isn't Live nor History -->
+    <ion-title *ngIf="currentPage != 'IndexLive' && currentPage != 'IndexHistory'" class="ion-hide-md-up" slot="start">
+      {{ service.currentPageTitle }}
+    </ion-title>
+    <!-- page title is hidden when 'smartphone resolution' is reached -->
+    <ion-title class="ion-hide-md-down" slot="start">
+      {{ service.currentPageTitle }}
+    </ion-title>
+    <ng-container *ngIf="service.currentEdge | async as edge">
+      <ng-container *ngIf="currentPage == 'IndexLive' || currentPage == 'IndexHistory'">
+        <ion-buttons>
+          <ion-segment mode="md" [(ngModel)]="currentPage" class="ion-justify-content-center">
+            <ion-segment-button (click)="updateLiveHistorySegment('IndexLive')" value="IndexLive">
+              <ion-label translate>General.live</ion-label>
+            </ion-segment-button>
+            <ion-segment-button (click)="updateLiveHistorySegment('IndexHistory')" value="IndexHistory">
+              <ion-label translate>General.history</ion-label>
+            </ion-segment-button>
+          </ion-segment>
+        </ion-buttons>
+      </ng-container>
+    </ng-container>
+    <!-- this title tag is for spacing purpose only -->
+    <ion-title class="ion-hide-md-down" slot="end">
+    </ion-title>
+    <ng-container *ngIf="service.currentEdge | async as edge">
+      <ion-buttons slot="end">
+        <ng-container *ngIf="edge.roleIsAtLeast('admin')">
+          <ng-container *ngIf="edge.currentData | async as currentData">
+            <ng-container *ngIf="currentData.summary.system as sum">
+              <ion-button (click)="presentSingleStatusModal()"
+                *ngIf="currentPage == 'IndexLive' || currentPage == 'IndexHistory' || currentPage == 'EdgeSettings'"
+                color="light" fill="solid" shape="round">
+                <ng-container [ngSwitch]="sum.state">
+                  <ion-icon *ngSwitchCase="0" color="success" name="checkmark-circle-outline"></ion-icon>
+                  <ion-icon *ngSwitchCase="1" color="success" name="information-outline"></ion-icon>
+                  <ion-icon *ngSwitchCase="2" color="warning" name="alert-outline"></ion-icon>
+                  <ion-icon *ngSwitchCase="3" color="danger" name="alert-outline"></ion-icon>
+                </ng-container>
+              </ion-button>
+            </ng-container>
+          </ng-container>
+          <ng-container *ngIf="env.debugMode && edge.roleIsAtLeast('owner')">
+            <ion-button (click)="isSystemLogEnabled = !isSystemLogEnabled">
+              <ion-icon name="bug-outline"></ion-icon>
+            </ion-button>
+          </ng-container>
+        </ng-container>
+      </ion-buttons>
+    </ng-container>
+  </ion-toolbar>
+
+  <ng-container *ngIf="currentPage == 'IndexHistory'">
+    <ion-toolbar color="secondary">
+      <ion-buttons class="ion-justify-content-center">
+        <pickdate></pickdate>
+      </ion-buttons>
+    </ion-toolbar>
+  </ng-container>
+</ion-header>
